KSENROLL-3907 Apply works and filtering can work on a single dump file
git-svn-id: https://svn.kuali.org/repos/student/tools/svn-history-tools/trunk@49580 0ac02d80-fb3f-11dc-a574-c57244a850eb

+++ scripts/copy-rename/filter-dump-files.sh
@@ -7,15 +7,22 @@ PIPE=/tmp/filter-dump-files.pipe
 DELETE=$1
 DEBUG=$2
 
-# delete any existing filtered dump files
-if test -n "$DELETE"
+# need to pass in the name of the rX-rY-join.dat file to use
+SPECIFIC=$3
+
+# only delete files if not in specific mode.
+if test -z "$SPECIFIC"
 then
+    # delete any existing filtered dump files
+    if test -n "$DELETE"
+    then
     
-    echo "rm -f *-filtered.dump"
-    rm -f *-filtered.dump
+        echo "rm -f *-filtered.dump"
+        rm -f *-filtered.dump
+
+    fi
 
 fi
-
 
 echo "rm -f \"$PIPE\""
 
@@ -41,27 +48,36 @@ then
 
 fi
 
-JOIN_WORK=$(ls -l *-join.dat | awk '{print $9}')
+if test -n "$SPECIFIC"
+then
 
-for WORK in $JOIN_WORK
-do
+    echo "Processing $SPECIFIC"
+    echo "$SPECIFIC" >> $PIPE
 
-    TARGET_REV=$(echo $WORK | cut -d- -f1)
+else
 
-    FILTERED_DUMP="$TARGET_REV-filtered.dump"
+    JOIN_WORK=$(ls -l *-join.dat | awk '{print $9}')
+
+    for WORK in $JOIN_WORK
+    do
+
+        TARGET_REV=$(echo $WORK | cut -d- -f1)
+
+        FILTERED_DUMP="$TARGET_REV-filtered.dump"
 
     
-    if test -f $FILTERED_DUMP
-    then
-        echo "$FILTERED_DUMP exists, skipping"
-        continue
-    fi
+        if test -f $FILTERED_DUMP
+        then
+            echo "$FILTERED_DUMP exists, skipping"
+            continue
+        fi
 
-    echo "Processing $WORK" 
-    echo "$WORK" >> $PIPE
+        echo "Processing $WORK" 
+        echo "$WORK" >> $PIPE
     
-done
+    done
 
+fi
 # hope to close the pipe
 cat /dev/null >> $PIPE
